Implement NumbaEngine
