+++ src/qibojit/__init__.py
@@ -1,1 +1,2 @@
 __version__ = "0.0.5.dev0"
+from qibojit.custom_operators.engines import NumbaEngine
